Fixed death animation loop and skull panel layout
1. I changed the animationRepeatCount to 1 for both golems in regards
to the death animation.

2. I originally preferred to have the skulls only show up on the panel
after 3 seconds, but stackview ruined those dreams. Because of the
stackview,
it made the placement of the skulls not appear in the slots,
but when all 3 skulls appear, the images looked fine.
To have a quick fix to that, I changed alpha to 0 and it worked.