Final App Store ready update...
Changed a couple UI Labels to keep in compliance with apple’s terms.

Submitted the app as it is and got a rejection:
5. 1.1 LEGAL: PRIVACY - DATA COLLECTION AND STORAGE
Legal - 5.1.1

We noticed that your app requires users to register with personal
information to access non account-based features. Apps cannot require
user registration prior to allowing access to app content and features
that are not associated specifically to the user.

We've attached screenshot(s) for your reference.

Next Steps

User registration that requires the sharing of personal information
must be optional or tied to account-specific functionality.
Additionally, the requested information must be relevant to the
features.